[src] Introduce DisposableObject and use it. (#13316)
Add a new DisposableObject class, which is a new base class below
NativeObject, to be used for types that aren't reference counted, but still
disposable. NativeObject is now implemented on top of this type, adding
support for reference counting.

Also change several type to use DisposableObject as their base type, which
makes it possible to share a lot more code and delete a good chunk of
duplicated code.