[Security] Fix call to AuthorizationExecuteWithPrivileges to pass a null-terminated array for the arguments. Fixes #13465. (#13472)
It looks like this was a latent bug that was uncovered due to unrelated code
changes. The problem is that the 'AuthorizationExecuteWithPrivileges' function
takes a null-terminated array as the arguments, and we were not
null-terminating the array we were passing (on purpose at least - it probably
have happened in the past by accident).

Fixes https://github.com/xamarin/xamarin-macios/issues/13465.