[Foundation] Remove conditional code that wasn't conditional for NSValueTransformer.
The conditional code was always true pre-NET. There's no reason to change it
for NET, so the condition can just be completely removed.

Clearly this wasn't the intention, and the condition was intended to be false
on macOS, but it's too late now to fix this accidental breaking change when it
occurred over 6 years ago.

Ref: https://github.com/xamarin/maccore/commit/3bc5c33bbc7c0eaf32ba7bb412b76c087b1365c0