[msbuild] Make the Codesign task able to sign files and directories that may depend on eachother.
Implement support for ordering signing so that directories containing files that
also must be signed are signed after those files.

This is implemented by:

1. Normalize all input (resolve symlinks, create full path, etc.)

2. Sort by path length (longest to shortest paths). This way we're certain that
   if we find a directory, we'll know that we won't find any files later in the list
   from inside that directory.

3. Group into an ordered list of buckets, where each bucket contains files and
   directories that don't depend on eachother (i.e. they can all be signed in parallel).

This makes it possible to call Codesign once, listing both the app bundle itself,
and all the individual files or directories inside that need signing, and the Codesign
task will sign the items in an order that ensures parent directories are always signed
after any files or directories inside.