[msbuild] Rework code signing to sign everything at the end.
Rework code signing to sign everything with a single call to the Codesign task in
the _CodesignAppBundle target for the executable project, instead of having multiple
calls to the Codesign task from multiple targets (and projects). This makes it easier
to reason about what's being signed, and it also makes it easier to add files to
the signing process.