[uikit] Use a code-first approach to app-level debugging options
Change `CheckForIllegalCrossThreadCalls` and `CheckForEventAndDelegateMismatches`
from fields to properties.

This has two main benefits:

1. Setting them inside code (and part of the templates) makes the
features easier to discover and easier to tweak the conditions to enable
or disable them. No extra arguments hidden in project files :)

2. This allows the linker to remove the properties (when disabled) since
they are not set from the static constructor (like the current fields)

Note: this is a draft, what's missing are

1. update all templates (ideally including MAUI ones in a different PR)

2. update the linker code to detect the properties (if never set) to
remove the feature. No more settings would be needed