[msbuild] Use a unique item group name for the app extension properties we'll serialize.
We have three tasks that executes for app extension signing:

* `_ReadAppExtensionCodesignProperties`: reads app extension signing info from
  disk (for the main app bundle).
* `_CodesignAppExtensions`: does the actual signing (during signing in the main
  app bundle), if there's anything to sign.
* `_PrepareCodesignAppExtension`: stores signing info to disk (for app
  extensions).

A problem arises if both of the following are true:

* These three targets use the same item group name for storing app extension
  signing info.
* We store signing info before reading it, like this order:
	* `_PrepareCodesignAppExtension`
	* `_ReadAppExtensionCodesignProperties`
	* `_CodesignAppExtensions`

In that case, the `_PrepareCodesignAppExtension` target will store the app
extension signing info in the same item group as
`_ReadAppExtensionCodesignProperties`, and then even if
`_ReadAppExtensionCodesignProperties` didn't find anything to read,
`_CodesignAppExtensions` will find something to sign (which is not what we
want).

Avoid this by using a different item group name in
`_PrepareCodesignAppExtension`.