[msbuild] Fixes the Windows Unzip task
Deleting the Extraction folder is not valid because it may contain unrelated files to the zip.

Since the ZipFile API requires the extraction path should not exist, we now use a temp folder, and then copy the extracted folders/files to the `ExtractionPath`.

This was breaking the Archive feature from VS, because the zip file is inside the extraction path, so it was deleted before trying to unzip it.