[foundation] add `NSMutableDictionary` helper to avoid creating extra `NSString` instances
That makes it easier on the caller, e.g. `CGContextPDF`, to avoid
creating short-lived `NSString` instance without requiring a lot
of code.

Downside: it becomes ambiguious if `NSString` is used, e.g.
`SecCertificate` for the compiler to decide between `NSObject` and
`string` overloads (but it's easy to fix and has no other impact).