[tools] Remove XAMCORE_4_0 condition in linker source.
Thinks are working fine as-is (with this XAMCORE_4_0 variable set to false),
and I see no particular reason in the code to change it, nor does the original
implementation explain much (b2bcad7a947d502db1355f017966a897966c9bb1).

So just remove this XAMCORE_4_0 condition as if it had never existed.