Implement a way to have non-UI thread finalization possible
Updated the NSObjectFlags to be a short, and that way we can fit new
values in at the code of possibly slightly increased memory usage,
but it might be covered by padding anyway.

Added NSObjectFlagsRequiresUIThread which is then used in the generator
to signal whether the UI thread is required for object finalization.

Also added a feature flag to enable this behaviour, while keeping the
old behaviour for compatibility.

Not sure what to do about user types, so it keeps the old behaviour for
now.