FSEventStream: support FSEventStreamCreateFlags.UseExtendedData
When UseExtendedData is specified, the event data type changes
from CFString to CFDictionary; the dictionary contains the path
and inode information for the file and may be extended in the future
with other fields. Previously this was crashing because we assumed
CFString always.

Adds file system tests which cover the existing (non-extended) and
fixed (extended) creation modes, along with using a dispatch queue
instead of run loop.

Fixes https://github.com/xamarin/xamarin-macios/issues/12007