FSEventStream: bind FSEventStreamCreateRelativeToDevice, FSEventStreamGetDeviceBeingWatched
Introduce FSEventStreamCreateOptions to avoid a slew of .ctor overrides,
and make it easier to specify DeviceToWatch and SinceWhenId. SinceWhenId
was previously only exposed on the "low level" .ctor, and it a rather
important parameter for supporting events that may have happened
while the application was not running.

Make the existing constructors wrap FSEventStreamCreateOptions to avoid
API break.