[CI] Build the rollback file in the build bot.
Or make files canculate the version of the nuget based on the hash of
the source, because VSTS does a merge of the branhc rather than checking
out, the version calculated in the tests bot via de makefiles is
different to the one that was used in the build bots.

We solve this by creating the rollback file at build time so that it is
consumed by the test bot.