[CI] Use public bots to run tests.
The internal bots are giving issues running xtro, public ones are ok.
There is no real reason to use the internal pool to run tests and we can
debug the bots in that pool better.

We move to run tests in the public bots unless told otherwise (when
using devices for example).