[Foundation] Fix numerous binding issues in NSAttributedString. Fixes #14489.
There were numerous differences in the managed API for NSAttributedString
between macOS and other platforms: most notably macOS would use 'out'
parameters, while the other platforms would use 'ref' parameters (possibly due
to this being very old API, and we didn't have proper support for 'out'
attributes when it was implemented).

This PR unifies the API between all platforms when it can (either now or in
XAMCORE_5_0).

It's hard to determine whether the changes don't break the API: api-diff is
happy locally, but the results on the bots will be the final determination.

Fixes https://github.com/xamarin/xamarin-macios/issues/14489.